:sparkles: Add colors to displacy automatically (#29)
* :sparkles: Add colors to displacy automatically

* :passport_control: Use model checkpoint upload on box to avoid google drive block

* :bug: Fix import
