+++ zshot/__init__.py
@@ -1,1 +1,2 @@
 from zshot.zshot import MentionsExtractor, Linker, Zshot, PipelineConfig  # noqa: F401
+from zshot.utils.displacy import displacy  # noqa: F401
