+++ zshot/linker/blink_linker.py
@@ -9,7 +9,7 @@ from appdata import AppDataPaths
 from spacy.tokens import Doc
 
 from zshot.linker.linker import Linker
-from zshot.utils import download_file
+from zshot.utils.utils import download_file
 
 MODELS_CACHE_PATH = os.getenv("BLINK_MODELS_CACHE_PATH") if "BLINK_MODELS_CACHE_PATH" in os.environ \
     else AppDataPaths(f"{Path(__file__).stem}").app_data_path + "/"
