+++ zshot/utils/__init__.py
@@ -1,0 +1,1 @@
+from zshot.utils.utils import download_file  # noqa: F401
