+++ zshot/utils/utils.py
@@ -3,8 +3,8 @@ import os
 import pathlib
 import shutil
 from urllib.request import urlopen
+
 import requests
-from spacy.tokens import Doc
 from tqdm.auto import tqdm
 
 
@@ -26,23 +26,3 @@ def download_file(url, output_dir="."):
         with tqdm.wrapattr(r.raw, "read", total=total_length, desc=f"Downloading {filename}") as raw:
             with path.open("wb") as output:
                 shutil.copyfileobj(raw, output)
-
-
-def ents_colors(doc: Doc):
-    """
-    Can be used to derive colors for entities in a Spacy document.
-    A color for each entity type in generated, using the entity label hash
-    :param doc: A Spacy document with entities
-    :return: A colors dictionary containing a color for each entity type
-    """
-
-    def color_from_label(label: str):
-        hash_s = hash(label)
-        r = (hash_s & 0xFF0000) >> 16
-        g = (hash_s & 0x00FF00) >> 8
-        b = hash_s & 0x0000FF
-        return '#%02x%02x%02x' % (r, g, b)
-
-    labels = set(ent.label_ for ent in doc.ents)
-    colors = dict([(ent, color_from_label(ent)) for ent in labels])
-    return colors
